Finish prototype
